added validation of user input

+++ sanitize.sh
@@ -1,18 +1,22 @@
 #!/bin/bash
-printf "---------------"
+printf "===============\n"
 printf "This script is potentially dangerous! \nIt WILL destroy data and make said data unrecoverable! \n \n"
+printf "===============\n"
 printf "Please choose disk to sanitize.\nBe mindful of which disk the OS is written to! \n"
-printf "---------------"
 lsblk
 read -p "Enter the disk name *EXACTLY* as shown in the chart above: " disktokill
 
 while true; do
 read -p "You selected $disktokill. Is this correct? (y/n)" yn
-    case $yn in
-        [Yy]* ) sudo dd if=/dev/zero of=/dev/$disktokill bs=64K status=progress ;
-               break;;
-        [Nn]* ) printf "Aborting sanitization!\n" ;
+    if [-f /dev/$disktokill]; then
+     case $yn in
+         [Yy]* ) sudo dd if=/dev/zero of=/dev/$disktokill bs=64K status=progress ;
                 break;;
-        * ) printf "Please answer yes or no." ;
-    esac
+         [Nn]* ) printf "Aborting sanitization!\n" ;
+                 break;;
+         * ) printf "Please answer yes or no." ;
+     esac
+    else
+     printf "$disktokill does not exist in /dev/ \n"
+    fi
 done
